+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Changelog
 
+
+## [3.0.0] - 2025-yy
+
+### Added
+- Added side view of the airspace situation. (#345)
+
+- Added option to use cabin pressure as static pressure, in order to compute
+  pressure altitude and vertical distance to airspaces (#546, #454)
+
+
 ## [2.34.6] - 2025-08-07
 
 ### Fixed
 
-- Fixed error where many data connections appear twice in the data connection dialog.
+- Fixed error where many data connections appear twice in the data connection
+  dialog.
 
 
 ## [2.34.5] - 2025-08-05